requirements.txt
pandas>=2.2
pyyaml>=6.0
requests>=2.31
pandera>=0.18
python-dateutil>=2.9
streamlit>=1.37
plotly>=5.24
